+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [//]: ### (Fixed)
 [//]: ### (Security)
 
+## [1.1.1] - 2016-03-26
+### Fixed
+- Error when loading user data [@negebauer]
+
+## [1.1.0] - 2016-03-26
 ### Added
 - Improve user data update
   - Verify ignore course acronyms separated by space [@negebauer]
@@ -43,7 +48,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Main app [@negebauer]
 - Yarn support [@mrpatiwi]
 
-[Unreleased]: https://github.com/open-source-uc/sincding/compare/v1.0.9...HEAD
+[Unreleased]: https://github.com/open-source-uc/sincding/compare/v1.1.0...HEAD
+[1.1.1]: https://github.com/open-source-uc/sincding/compare/v1.0.0...v1.1.1
+[1.1.0]: https://github.com/open-source-uc/sincding/compare/v1.0.9...v1.1.0
 [1.0.9]: https://github.com/open-source-uc/sincding/compare/v1.0.8...v1.0.9
 [1.0.8]: https://github.com/open-source-uc/sincding/compare/v1.0.7...v1.0.8
 [1.0.7]: https://github.com/open-source-uc/sincding/compare/v1.0.6...v1.0.7